Moved all "view" components into view folder
