+++ src/app/components/views/help/help.component.ts
