+++ src/app/components/views/qc-timer/qc-timer.component.spec.ts
