+++ src/app/components/views/ticket-generator/ticket-generator.component.ts
